requirements.txt
langchain
langchain-openai
langchain-community
openai
python-dotenv
pypdf2